fastapi_backend/requirements.txt
# ===============================
# FASTAPI USER PANEL
# ===============================

fastapi==0.110.0
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.25
pymysql==1.1.0

# Authentication
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9

# Validation
pydantic==2.6.1
pydantic[email]==2.6.1

# Background Tasks (Bonus)
celery==5.3.6
redis==5.0.1

# Environment Config
python-dotenv==1.0.1

passlib==1.7.4